# src/resources/swagger.yaml
openapi: 3.0.1
info:
  title: EWORLD-BETA
  description: ''
  contact:
    email: jerryle.vn@gmail.com
  version: 0.0.1
servers:
- url: http://localhost:3001/api
- url: http://petstore.swagger.io/v2
...
paths:
...
  /auth/register-otp/:
    post:
      tags:
      - auth
      summary: Validate mã OTP để đăng ký tài khoản
      operationId: registerOtp
      requestBody:
        description: Nội dung
        content:
          application/json:
            schema:
              required:
              - phone
              - code
              type: object
              properties:
                phone:
                  type: string
                code:
                  type: string
            example:
              phone: '0975234889'
              code: '0050'
        required: true
      responses:
        200:
          description: Không trùng SĐT và validate code OTP thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYxYmFiMzliN2JmODdmYmMwNzg4ODBkYiIsImlhdCI6MTYzOTYyNTYyNywiZXhwIjoxNjcxMTYxNjI3fQ.J3lSgt04G4x-r5us7RtRHpV3Mx9auVCaMWxHrRpX88U
                  userId: 61bab39b7bf87fbc078880db
                  account:
                    profile:
                      fullName: Check
                    phone: '+84888253236'
                    type: customer
                    authentication:
                      lock: false
                      status: false
                      isPhoneVerified: true
                    _id: 61bab39b7bf87fbc078880db
                    createdAt: '2021-12-16T03:33:47.780Z'
                    updatedAt: '2021-12-16T03:33:47.780Z'
                status: 200
                message: Register successfully
                errors: null
        400:
          description: Không thành công vì trùng số điện thoại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                data: null
                status: 400
                message: Code is not valid
                errors: null
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        data:
          oneOf:
            - type: object
              nullable: true
            - type: array
              nullable: true
              items:
                type: object
        status:
          type: integer
          format: int32
        errors:
          nullable: true
          oneOf:
            - type: object
              nullable: true
            - type: array
              nullable: true
              items:
                type: object
        message:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT